validation-service/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1.13
djongo==1.3.6
dnspython==2.4.2
Flask==3.0.0
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
pika==1.3.2
pip==23.3.1
prompt-toolkit==3.0.41
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==68.0.0
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wheel==0.41.2
